+++ src/sql/src/plan/statement/dml.rs
@@ -257,13 +257,13 @@ pub fn describe_tail(
     let progress = options.progress.unwrap_or(false);
     const MAX_U64_DIGITS: u8 = 20;
     let mut desc = RelationDesc::empty().with_named_column(
-        "timestamp",
+        "mz_timestamp",
         ScalarType::Decimal(MAX_U64_DIGITS, 0).nullable(false),
     );
     if progress {
-        desc = desc.with_named_column("progressed", ScalarType::Bool.nullable(false));
+        desc = desc.with_named_column("mz_progressed", ScalarType::Bool.nullable(false));
     }
-    desc = desc.with_named_column("diff", ScalarType::Int64.nullable(true));
+    desc = desc.with_named_column("mz_diff", ScalarType::Int64.nullable(true));
     for (name, ty) in sql_object.desc()?.iter() {
         let mut ty = ty.clone();
         if progress {